+++ astro/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-core/astro
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.14.1
+
 ## 1.3.1
 
 ### Patch Changes