+++ testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # testing
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.14.1
+  - @bs-plugins/jira@1.1.1
+  - @bs-plugins/template@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes